pHDL: My HDL interop library
I'm sick of all these silly scripts that don't actually parse input files
correctly but instead just require a bunch of manual fixups to make our flow
work.  The point of this library is to be a generic interface to RTL written in
a HDL that tools can use to reason about designs.